+++ src/add_label_to_query.ts
@@ -13,7 +13,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import { parser, VectorSelector } from '@prometheus-io/lezer-promql';
+import { parser, VectorSelector } from 'lezer-metricsql';
 
 import { PromQueryModeller } from './querybuilder/PromQueryModeller';
 import { buildVisualQueryFromString } from './querybuilder/parsing';