+++ src/querybuilder/metricsql-functions/aggregations/index.ts
@@ -1,0 +1,11 @@
+import { QueryBuilderOperationDef } from "../../shared/types";
+
+import { getOperationFunctions } from "./operations";
+import { getOverTimeFunctions } from "./overTime";
+
+export function getAggregations(): QueryBuilderOperationDef[] {
+  return [
+    ...getOverTimeFunctions(),
+    ...getOperationFunctions(),
+  ]
+}